Добавлена модульность (aurusov/rdo-xtext#55)
